Changed text color to white
